Revert "第一形态人&党内谈话-审核修改 (#178)"
This reverts commit 35ca1d7179ea231d51485c6dfce21d5e391a2e80.

+++ jeecg-boot-module-system/src/main/java/org/jeecg/modules/system/controller/SysTenantController.java
@@ -118,7 +118,7 @@ public class SysTenantController {
             // 过滤掉已被引用的租户
             List<String> idList = new ArrayList<>();
             for (String id : ls) {
-                Long userCount = sysTenantService.countUserLinkTenant(id);
+                int userCount = sysTenantService.countUserLinkTenant(id);
                 if (userCount == 0) {
                     idList.add(id);
                 }
